Bug fixed. Argumento de setDimensao foi mudado para double.

+++ src/com/grupo/device/SmartBulb.java
@@ -34,7 +34,7 @@ public class SmartBulb extends SmartDevice{
 
     }
 
-    protected void setDimensao(int dimensao){
+    protected void setDimensao(double dimensao){
 
     }
